+++ src/block/block_manager.cc
@@ -559,7 +559,8 @@ int BlockManager::AddNewBlock(
 
     AccountManager::Instance()->AddBlockItemToDb(block_item, db_batch, is_kv_sync);
     ShardStatistic::Instance()->AddStatistic(block_item);
-    if (common::GlobalInfo::Instance()->data_service_node_for_net_id() == block_item->network_id()) {
+    if (common::GlobalInfo::Instance()->data_service_node_for_net_id() ==
+            (int32_t)block_item->network_id()) {
         ck_client_->AddNewBlock(block_item);
     }
